requirements.txt
chromadb
sentence-transformers
numpy
pypdf
python-docx
tqdm